changes in wrapper.py
