+++ wrapper/kerasclassifier.py
@@ -1,30 +1,5 @@
-class Model(object):
-    def __init__(self, name, approach):
-        self.name = name
-        self.approach = approach
-
-    def predict(self):
-        pass
-
-
-class NonDeterministicMethods(Model):
-    def save_model(self):
-        pass
-
-    def load_model(self):
-        pass
-
-
-class MachineLearning(NonDeterministicMethods):
-    pass
-
-
-class Supervised(MachineLearning):
-    pass
-
-
-class Classifier(Supervised):
-    pass
+from .inheritance import Model
+from .inheritance import Classifier
 
 
 class KerasClassifier(Classifier):
